refactor: change the structure of composition plugin
This makes for easier editing.

Also move some things around in core.
