+++ @remirror/core-extensions/src/extensions/index.ts
@@ -2,3 +2,4 @@ export * from './placeholder';
 export * from './history';
 export * from './composition';
 export * from './inline-cursor-target';
+export * from './types';
