--- @remirror/core-extensions/src/extensions/position/cursor.ts
@@ -1,9 +1,0 @@
-import { Extension } from '@remirror/core';
-
-export interface CursorPositionOptions {}
-
-export class CursorPosition extends Extension<CursorPositionOptions> {
-  get name(): 'cursorPosition' {
-    return 'cursorPosition';
-  }
-}
