--- @remirror/core-extensions/src/extensions/position/index.ts
@@ -1,2 +1,0 @@
-export * from './cursor';
-export * from './menu';
