--- @remirror/core-extensions/src/extensions/position/menu.ts
@@ -1,11 +1,0 @@
-import { Extension } from '@remirror/core';
-
-export interface MenuPositionOptions {
-  holder?: string;
-}
-
-export class MenuPosition extends Extension<MenuPositionOptions> {
-  get name(): 'menu' {
-    return 'menu';
-  }
-}
