--- @remirror/core-extensions/src/extensions/position/state.ts
@@ -1,1 +1,0 @@
-export class CursorPositionState {}
