+++ @remirror/core/src/helpers/coords-at-pos.ts
@@ -1,4 +1,6 @@
-import { Cast, EditorView, isElementNode, isTextNode } from '@remirror/core';
+import { EditorView } from '../types';
+import { Cast } from './base';
+import { isElementNode, isTextNode } from './document';
 
 export const textRange = (node: Node, from?: number, to?: number) => {
   const range = document.createRange();
@@ -13,7 +15,7 @@ function singleRect(object: Range | Element, bias?: number) {
 }
 
 /**
- * EditorView.cordsAtPos returns wrong coordinates if we are at a linebreak.
+ * EditorView.cordsAtPos returns wrong coordinates when at a linebreak.
  * @see https://github.com/scrumpy/tiptap/pull/228
  */
 export const coordsAtPos = (view: EditorView, pos: number, end = false) => {
