+++ @remirror/core/src/index.ts
@@ -5,9 +5,8 @@ export * from './mark-extension';
 export * from './types';
 export * from './commands';
 export * from './nodes';
-export * from './document-helpers';
-export * from './helpers';
 export * from './utils';
 export * from './portal-container';
 export * from './constants';
 export * from './selection';
+export * from './helpers';
