+++ @remirror/core/src/selection/gap-cursor.ts
@@ -1,7 +1,7 @@
 import { ResolvedPos, Schema, Slice } from 'prosemirror-model';
 import { Selection, SelectionBookmark } from 'prosemirror-state';
 import { GAP_CURSOR_IGNORED_NODE, Side } from '../constants';
-import { nodeNameMatchesList } from '../document-helpers';
+import { nodeNameMatchesList } from '../helpers/document';
 import { Mapping } from '../types/aliases';
 import { EditorSchema, NodeMatch, PMNode } from '../types/base';
